text2sql/requirements.txt
semantic-kernel==1.4.0
azure-search
azure-search-documents==11.6.0b3
aioodbc
azure-identity
